Organize scripts in bin/command #65 (#66)
Changelog
====
- Adds `--diff-filter=d` option to `git diff` in
`bin/command/check-diff-format`, to ignore deleted files. Previously it
tried to check the code difference of the deleted files and it was an
unexpected behavior.
- Fixes `bin/command/one-time-setup` to pass the formatting CI/CD
pipeline.
- Deletes `bin/command/kill-apps` since it is not in use anymore.
- Merges `bin/command/flask-server-setup`, `bin/command/idealog-backup`
into `bin/command/manage-server`.

Detailed Description
====
This PR is requested to organize the script files in bin/command, since
there are too many script files unnecessarily. We can remove scripts not
in use and merge some scripts, and it will make it easy to understand
the scripts in the repo.